.history/backend/requirements_20260421172521.txt
Flask>=3.0,<4.0
flask-cors>=4.0,<5.0
PyJWT>=2.8,<3.0
bcrypt>=4.1,<6.0
requests>=2.31,<3.0
scikit-learn>=1.5,<2.0
numpy>=2.0,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
python-whois>=0.9,<1.0

# Optional MongoDB support (not required for the current SQLite backend)
pymongo>=4.8,<5.0
